Fix error in awg writing in lib

+++ src/lib/rp-daq-lib.h
@@ -63,7 +63,7 @@ extern volatile uint32_t *slcr, *axi_hp0;
 extern void *pdm_sts, *reset_sts, *cfg, *ram, *buf, *dio_sts;
 extern uint16_t *pdm_cfg;
 extern uint64_t *adc_sts, *dac_cfg; 
-extern int16_t *awg_0_cfg, *awg_1_cfg;
+extern uint32_t *awg_0_cfg, *awg_1_cfg;
 
 // init routines
 extern uint32_t getFPGAId();
